refactor: header frame 분리 및 Main topic 생성
- wsceactor에서도 사용하는 header frame을 공용 컴포넌트로 분리
- Main 페이지에 사용되는 컴포넌트 Main 토픽으로 이동
