+++ src/frontend/web/src/components/Common/MainHeaderFrame.ts
@@ -1,0 +1,11 @@
+import React, { ReactChild, ReactChildren } from 'react';
+import styled from 'styled-components';
+
+const MainHeaderFrame = styled.header`
+  display: flex;
+  justify-content: space-between;
+  background-color: beige;
+  padding: 6px 0px;
+`;
+
+export default MainHeaderFrame;
