+++ src/frontend/web/src/components/Main/Chat/Header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import expandIcon from '../../icon/expand.svg';
+import expandIcon from '../../../icon/expand.svg';
 
 const ChannelHeader = styled.article`
   display: flex;
